added macro to mark all symbols that are to be exported to a DLL (MSVC)
This is necessary, because on windows the linker will optimize out all unused, unmarked symbols. Only symbols with declspec(dllexport) will survive in the DLL.
And dllexport must only be set when actually building the DLL. When just using the headers, you'll need to leave it out (or use dllimport as an optimization).