added some more required DLL exports for external trackers
Note the disabled warning in TrackedObject. Otherwise, the whole STL container needs to be exported to the DLL, which is very problematic for a std::map.
Just disabling the warning works in this case, because the member is private and can not be accessed directly from outside of the DLL anyway.

For more info on this, see https://support.microsoft.com/en-us/kb/168958