added support for user annotations
Annotations are visual markers (such as arrows, labels, ...), tied to frames of a video. The user can create new markers (A->arrow, L->label), move their endpoints (handles) around and delete a selected annotation with DEL.
Annotations will be automatically serialized to a .csv file in the same folder of the input file (and with the suffix ".annotations.csv"). This csv file can likely be easily read in R or Python.
Serialized annotations will be reloaded when a video is opened again. This allows for least required user-interaction. Yet an issue might be that you can not "abort" or "reset" annotations after having edited them. Might need UI for that (resetting is really easy, code-wise).