Restructured code, CSV out changed, performance fix?
Some code restructuring I wanted to do for a long time. Also improved on the exporters. This included changing from steady clock timestamp to system ones.
The performance issue was apparently due to the use of size() and .at() for each and every QList access in the trajectories. .at() on the last element was replaced by .back() and the size is now being counted.
Also the number of valid children is now being counted and only recomputed when necessary. As this is a rare event (user triggered) this should boost performance quite a bit.
see https://github.com/BioroboticsLab/biotracker_core/issues/144